+++ lib/platform_utils.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/foundation.dart';
+
+bool get isDesktopPlatform =>
+    defaultTargetPlatform == TargetPlatform.linux ||
+        defaultTargetPlatform == TargetPlatform.windows ||
+        defaultTargetPlatform == TargetPlatform.macOS;